% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFires-v1.R, R/buildFires.R
\name{buildFires}
\alias{buildFires}
\title{Building fire list}
\usage{
buildFires(fileFiresFIraw, fileFiresBIraw, method = "fast")

buildFires(fileFiresFIraw, fileFiresBIraw, method = "fast")
}
\arguments{
\item{fileFiresFIraw}{Path of file containing Fire Intensity}

\item{fileFiresBIraw}{Path of file containing Iterations}

\item{method}{Mode to adopt when importing data: "fast" is quicker bur uses more memory, "slow" is the opposite}
}
\value{
Fire list of Fire Intensity for each Iteration

Fire list of Fire Intensity for each Iteration
}
\description{
Build fire list from Burn-P3 output storing Fire Intensity and Iterations

Build fire list from Burn-P3 output storing Fire Intensity and Iterations
}
\examples{
fi <- buildFires(fileFiresFIraw, fileFiresBIraw)
fi <- buildFires(fileFiresFIraw, fileFiresBIraw, method = "fast")
}
