% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLR.R
\name{summaryLR}
\alias{summaryLR}
\title{Summary Loss Rate}
\usage{
summaryLR(final)
}
\arguments{
\item{final}{Loss Rate object produced with computeLR function}
}
\description{
Print summary of a Loss Rate object produced with computeLR function
}
\examples{
summaryLR(final)
}
