% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatisticsLR.R
\name{summaryStatisticsLR}
\alias{summaryStatisticsLR}
\title{Summary Statistics for Loss Rate}
\usage{
summaryStatisticsLR(final, damagedThreshold = 50)
}
\arguments{
\item{final}{Loss Rate object produced with computeLR function}

\item{damagedThreshold}{Loss Rate threshold for considering a structure damaged, defauilt is 50}
}
\value{
sf object with average LR, median LR, min LR, max LR, times exposed and times damaged for each impacted Pixel With Structure
}
\description{
Compute summary statistics of Loss Rate object produced with computeLR function
}
\examples{
summary <- summaryStatisticsLR(final)
}
