% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFires.R
\name{buildFires}
\alias{buildFires}
\title{Building fire list}
\usage{
buildFires(fileFiresFIraw, fileFiresBIraw)
}
\arguments{
\item{fileFiresFIraw}{Path of file containing Fire Intensity}

\item{fileFiresBIraw}{Path of file containing Iterations}
}
\value{
Fire list of Fire Intensity for each Iteration
}
\description{
Build fire list from Burn-P3 output storing Fire Intensity and Iterations
}
\examples{
fi <- buildFires(fileFiresFIraw, fileFiresBIraw)
}
