% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructureGrid.R
\name{createStructureGrid}
\alias{createStructureGrid}
\title{Create structure grid}
\usage{
createStructureGrid(
  structures,
  cellSize = 100,
  mode = "maximal",
  internalBuffer = F,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{structures}{File containing structures' polygons}

\item{cellSize}{Length of edge of grid squares, in meters}

\item{mode}{Mode to adopt when counting structures in multiple squares: "maximal" to count each occurrence, "minimal" to count it just in square covering most of its surface}

\item{internalBuffer}{Distance from border of grid containing structures to be excluded, in meters. Default value is FALSE}

\item{xmin, ymin}{Specify a corner of portion of structures' file to consider}

\item{xmax, ymax}{Specify opposite corner of portion of structures' file to consider}
}
\value{
Grid of squares with edge cellSize, with count of associated structures
}
\description{
Create a structure grid of squares and count the number of structures in each one
}
\examples{
grid <- createBuildingGrid <- function(structures, 100, "maximal")
}
